Make vx32 to work when code loaded from shared library
On 64 bit architecutures vx32 assumes the code is in lower 4 gigs.
This is invalid with ASLR (-pie) or when loading vx32 as a dynamic
library object. This patch fixes that behaviour by copying over
the critical rts code to lower 4 gigs, and by adding one more
indirection level when calling vxrun_return, to allow for full
8 byte indexing.